define normal direction as general dynamics
