+++ tests/user_examples/extra_src/shared/eulerian_fluid_dynamics/eulerian_fluid_dynamics.h
@@ -193,10 +193,10 @@ class NonReflectiveBoundaryCorrection : public LocalDynamics, public DataDelegat
     Vecd vel_farfield_;
     StdLargeVec<Real> &rho_, &p_, &Vol_;
     StdLargeVec<Vecd> &vel_, &mom_, &pos_;
-    StdLargeVec<Vecd> n_;
     StdLargeVec<Real> inner_weight_summation_, rho_average_, vel_normal_average_;
     StdLargeVec<Vecd> vel_tangential_average_, vel_average_;
     StdLargeVec<int> &indicator_, smeared_surface_;
+    StdLargeVec<Vecd> &n_;
 };
 } // namespace fluid_dynamics
 } // namespace SPH
